chore: state for uploads
